requirements.txt
nicegui
pinocchio
numpy>=1.24.0
python-can
asyncio
pillow
meshcat
serial
pin
pin-pink
scipy>=1.8
qpsolvers>=1.6